fix: spacings and paddings
