+++ src/components/common/buttons/FlatButton.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { Button } from './Button';
+
+const FlatButton = styled(Button)`
+  background: transparent;
+  border: none;
+`;
+
+export {
+  FlatButton
+};
