+++ src/components/common/buttons/index.jsx
@@ -1,0 +1,5 @@
+export { Button } from './Button';
+export { ButtonBack } from './ButtonBack';
+export { ButtonDownload } from './ButtonDownload';
+export { FlatButton } from './FlatButton';
+export { PreviewButtonNext, PreviewButtonPrevious } from './PreviewButtons';
